tests: overlay refactor
